Update apt package list
Hopefully addresses #3.

`apt update` downloads a fresh list of available packages.
Running `apt install` with outdated list may fail with `404  Not Found`
as the local list points to a version which is no longer available.

(The tests with Dockerfile-* didn't catch this because the Dockerfiles
themselves do `apt-get update` early in order to install `sudo`.
But ALL.sh didn't :-)